[NO-ISSUE] args: better api for loading folders
The problem was that the tool was only searching for markdown files at
the first level and that the filter was very basic.

Now it has the ability to perform recursive searches for markdown files
and the user can pass in a regex filter to choose which markdown files
are going to get executed.