[NO-ISSUE] args: positional argument for files
Add a positional arg for the directory to load instead of specifying
`-m`. This makes the tool more inline with other classical linux tools
such as find.

Add a customized help message to categorize the options to help new
users finding their way in the tool.

The user can pass in a path or a file as the positional argument, which
makes testing a single file way more easy than before.